PIM-4395: re-organize mass edit processors
