+++ src/Pim/Bundle/EnrichBundle/Connector/Item/MassEdit/TemporaryFileCleaner.php
@@ -15,7 +15,7 @@ use Akeneo\Bundle\BatchBundle\Step\StepExecutionAwareInterface;
  * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class MassEditTemporaryFileCleaner extends AbstractConfigurableStepElement implements StepExecutionAwareInterface
+class TemporaryFileCleaner extends AbstractConfigurableStepElement implements StepExecutionAwareInterface
 {
     /** @var StepExecution */
     protected $stepExecution;
