+++ src/Pim/Bundle/EnrichBundle/Connector/Item/MassEdit/VariantGroupCleaner.php
@@ -28,7 +28,7 @@ use Symfony\Component\Validator\ConstraintViolationListInterface;
  * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class MassEditVariantGroupCleaner extends AbstractConfigurableStepElement implements StepExecutionAwareInterface
+class VariantGroupCleaner extends AbstractConfigurableStepElement implements StepExecutionAwareInterface
 {
     /** @var StepExecution */
     protected $stepExecution;
