+++ src/Pim/Bundle/EnrichBundle/Resources/config/connector/cleaners.yml
@@ -1,6 +1,6 @@
 parameters:
-    pim_enrich.mass_edit.temporary_file.cleaner.class: Pim\Bundle\EnrichBundle\Connector\Item\MassEdit\MassEditTemporaryFileCleaner
-    pim_enrich.mass_edit.add_to_variant.cleaner.class: Pim\Bundle\EnrichBundle\Connector\Item\MassEdit\MassEditVariantGroupCleaner
+    pim_enrich.mass_edit.temporary_file.cleaner.class: Pim\Bundle\EnrichBundle\Connector\Item\MassEdit\TemporaryFileCleaner
+    pim_enrich.mass_edit.add_to_variant.cleaner.class: Pim\Bundle\EnrichBundle\Connector\Item\MassEdit\VariantGroupCleaner
 
 services:
     pim_enrich.mass_edit.temporary_file.cleaner:
